refactor: extract resolveSlackToken from share.ts into shared slack helpers (#30407)
